+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.UnmodifiableObservableList;
 import seedu.address.model.task.ReadOnlyTask;
-import seedu.address.model.task.UniqueTaskList.TaskNotFoundException;
+import seedu.address.model.task.TaskNotFoundException;
 
 /**
  * Deletes a task identified using it's last displayed index from the task book.
