+++ src/main/java/seedu/address/model/task/TaskNotFoundException.java
@@ -1,0 +1,9 @@
+package seedu.address.model.task;
+
+/**
+ * Signals that an operation targeting a specified task in the list would fail because
+ * there is no such matching task in the list.
+ */
+public class TaskNotFoundException extends Exception {
+
+}
