+++ src/main/java/seedu/address/model/task/UniqueTaskList.java
@@ -26,12 +26,6 @@ public class UniqueTaskList implements Iterable<Task> {
         }
     }
 
-    /**
-     * Signals that an operation targeting a specified task in the list would fail because
-     * there is no such matching task in the list.
-     */
-    public static class TaskNotFoundException extends Exception {}
-
     private final ObservableList<Task> internalList = FXCollections.observableArrayList();
 
     /**
